使用预备金初步搭建

+++ ios_zhongyouzhongchou/ios_zhongyouzhongchou/classes/ZYZCMine(我的)/MyWallet(钱包)/WalletHome/Model/WalletHeadModel.h
@@ -11,5 +11,5 @@
 @interface WalletHeadModel : NSObject
 
 @property (nonatomic, assign) CGFloat cash;
-@property (nonatomic, assign) CGFloat uCash;
+@property (nonatomic, assign) NSInteger uCash;
 @end
